Rename 'rocks' package to 'database'
