+++ server/TypeDBServer.java
@@ -29,8 +29,8 @@ import com.vaticle.typedb.core.common.parameters.Options;
 import com.vaticle.typedb.core.concurrent.executor.Executors;
 import com.vaticle.typedb.core.migrator.MigratorClient;
 import com.vaticle.typedb.core.migrator.MigratorService;
-import com.vaticle.typedb.core.rocks.Factory;
-import com.vaticle.typedb.core.rocks.RocksFactory;
+import com.vaticle.typedb.core.database.Factory;
+import com.vaticle.typedb.core.database.RocksFactory;
 import com.vaticle.typedb.core.server.common.Command;
 import com.vaticle.typedb.core.server.common.CommandLine;
 import com.vaticle.typedb.core.server.common.Configuration;
