+++ test/behaviour/connection/ConnectionSteps.java
@@ -20,8 +20,8 @@ package com.vaticle.typedb.core.test.behaviour.connection;
 
 import com.vaticle.typedb.core.TypeDB;
 import com.vaticle.typedb.core.common.parameters.Options;
-import com.vaticle.typedb.core.rocks.RocksDatabase;
-import com.vaticle.typedb.core.rocks.RocksDatabaseManager;
+import com.vaticle.typedb.core.database.RocksDatabase;
+import com.vaticle.typedb.core.database.RocksDatabaseManager;
 import io.cucumber.java.After;
 import io.cucumber.java.Before;
 import io.cucumber.java.en.Given;
