+++ test/behaviour/reasoner/verification/test/CorrectnessVerifierTest.java
@@ -21,9 +21,9 @@ package com.vaticle.typedb.core.test.behaviour.reasoner.verification.test;
 import com.vaticle.typedb.core.TypeDB;
 import com.vaticle.typedb.core.common.parameters.Arguments;
 import com.vaticle.typedb.core.common.parameters.Options;
-import com.vaticle.typedb.core.rocks.RocksDatabaseManager;
-import com.vaticle.typedb.core.rocks.RocksSession;
-import com.vaticle.typedb.core.rocks.RocksTransaction;
+import com.vaticle.typedb.core.database.RocksDatabaseManager;
+import com.vaticle.typedb.core.database.RocksSession;
+import com.vaticle.typedb.core.database.RocksTransaction;
 import com.vaticle.typedb.core.test.behaviour.reasoner.verification.CorrectnessVerifier;
 import com.vaticle.typedb.core.test.behaviour.reasoner.verification.CorrectnessVerifier.CompletenessException;
 import com.vaticle.typedb.core.test.behaviour.reasoner.verification.CorrectnessVerifier.SoundnessException;
