+++ test/integration/StringAttributeTest.java
@@ -27,7 +27,7 @@ import com.vaticle.typedb.core.common.parameters.Options;
 import com.vaticle.typedb.core.concept.answer.ConceptMap;
 import com.vaticle.typedb.core.concept.thing.Attribute;
 import com.vaticle.typedb.core.concept.type.AttributeType;
-import com.vaticle.typedb.core.rocks.RocksDatabaseManager;
+import com.vaticle.typedb.core.database.RocksDatabaseManager;
 import com.vaticle.typedb.core.test.integration.util.Util;
 import com.vaticle.typeql.lang.TypeQL;
 import org.junit.Test;
