+++ test/integration/logic/BUILD
@@ -30,7 +30,7 @@ host_compatible_java_test(
         "//concept:concept",
         "//logic:logic",
         "//pattern:pattern",
-        "//rocks:rocks",
+        "//database:database",
         "//traversal:traversal",
     ],
     test_class = "com.vaticle.typedb.core.logic.RuleTest",
@@ -50,7 +50,7 @@ host_compatible_java_test(
     test_class = "com.vaticle.typedb.core.logic.TypeInferenceTest",
     native_libraries_deps = [
         "//:typedb",
-        "//rocks:rocks",
+        "//database:database",
         "//common",
         "//common/test:util",
         "//concept:concept",
@@ -74,7 +74,7 @@ host_compatible_java_test(
     srcs = ["resolvable/UnifyRelationConcludableTest.java", "resolvable/Util.java"],
     native_libraries_deps = [
         # Internal dependencies
-        "//rocks:rocks",
+        "//database:database",
         "//:typedb",
         "//common:common",
         "//concept:concept",
@@ -98,7 +98,7 @@ host_compatible_java_test(
     native_libraries_deps = [
         # Internal dependencies
         "//:typedb",
-        "//rocks:rocks",
+        "//database:database",
         "//common:common",
         "//concept:concept",
         "//logic:logic",
@@ -121,7 +121,7 @@ host_compatible_java_test(
     native_libraries_deps = [
         # Internal dependencies
         "//:typedb",
-        "//rocks:rocks",
+        "//database:database",
         "//common:common",
         "//concept:concept",
         "//logic:logic",
@@ -145,7 +145,7 @@ host_compatible_java_test(
         # Internal dependencies
         "//:typedb",
         "//common:common",
-        "//rocks:rocks",
+        "//database:database",
         "//concept:concept",
         "//logic:logic",
         "//pattern:pattern",
