+++ test/integration/reasoner/BUILD
@@ -24,7 +24,7 @@ host_compatible_java_test(
     test_class = "com.vaticle.typedb.core.reasoner.ReasonerTest",
     native_libraries_deps = [
         "//:typedb",
-        "//rocks:rocks",
+        "//database:database",
         "//common:common",
         "//concurrent:concurrent",
         "//logic:logic",
@@ -46,7 +46,7 @@ host_compatible_java_test(
     test_class = "com.vaticle.typedb.core.reasoner.ExplanationTest",
     native_libraries_deps = [
         "//:typedb",
-        "//rocks:rocks",
+        "//database:database",
         "//concept:concept",
         "//common:common",
         "//concurrent:concurrent",
@@ -71,7 +71,7 @@ host_compatible_java_test(
     native_libraries_deps = [
         # Internal dependencies
         "//:typedb",
-        "//rocks:rocks",
+        "//database:database",
         "//common:common",
         "//concept:concept",
         "//concurrent:concurrent",
@@ -96,7 +96,7 @@ host_compatible_java_test(
     native_libraries_deps = [
         # Internal dependencies
         "//:typedb",
-        "//rocks:rocks",
+        "//database:database",
         "//common:common",
         "//concurrent:concurrent",
         "//concept:concept",
@@ -121,7 +121,7 @@ host_compatible_java_test(
     test_class = "com.vaticle.typedb.core.reasoner.resolution.PlannerTest",
     native_libraries_deps = [
         # Internal dependencies
-        "//rocks:rocks",
+        "//database:database",
         "//:typedb",
         "//common:common",
         "//concept:concept",
