+++ test/integration/rocks/DatabaseTest.java
@@ -16,11 +16,10 @@
  *
  */
 
-package com.vaticle.typedb.core.rocks;
+package com.vaticle.typedb.core.database;
 
 import com.vaticle.typedb.core.common.parameters.Options;
 import com.vaticle.typedb.core.graph.common.Encoding;
-import com.vaticle.typedb.core.test.integration.util.Util;
 import org.junit.Test;
 
 import java.io.IOException;
