+++ traversal/BUILD
@@ -24,7 +24,7 @@ package(
         "//pattern:__subpackages__",
         "//query:__subpackages__",
         "//reasoner:__subpackages__",
-        "//rocks:__pkg__",
+        "//database:__pkg__",
     ]
 )
